Locking and Centering Initial
Adding a notebook for a barebones implementation for the final target locking and centering method. 
some notes; when you run the program, the robot will spin to the left, until it detects a human. it would then theoretically call  a center method, which i have not yet written that would then take the x coordinate of the bounding box and center in the lens of the robot. while it isn't implemented, the robot just turns two iterations to the left, a simplistic substitute for a much more complex solution. 

the robot will turn until it detects the final target, which i have hard coded to be a human for now, 10 times. i wrote some comments regarding the camera.observe method in the last cell of the notebook. 

i will finish up the module by the lastest tuesday evening.