changes to optimize for following.
robot now turns to the right in smaller increments at a faster speed. 
the robot naturally moves to the left, so turning right will make it easier to lock onto the target after losing it. 

the following works well at distances less than around 10 feet, but more than that, it is much harder for the robot to center the object, due to its being smaller in the vision of the robot. should implement smaller "small increments" depending on the distance away from the target or something to help combat this issue